# Stop unwrapping the IF in the AGG(IF()) rewrite