--- a/presto-main/src/main/java/com/facebook/presto/sql/analyzer/FeaturesConfig.java
+++ b/presto-main/src/main/java/com/facebook/presto/sql/analyzer/FeaturesConfig.java
@@ -206,7 +206,7 @@ public class FeaturesConfig
     private boolean materializedViewDataConsistencyEnabled = true;
 
     private boolean queryOptimizationWithMaterializedViewEnabled;
-    private boolean aggregationIfToFilterRewriteEnabled = true;
+    private boolean aggregationIfToFilterRewriteEnabled;
 
     public enum PartitioningPrecisionStrategy
     {