Disabled the rule of AGG(IF()) rewrite by default. Update it to rewrite
    AGG(IF(condition, expr))
to
    AGG(IF(condition, expr)) FILTER (WHERE condition).

Stop unwrapping the IF expression in the aggregate, because the
true branch might return errors for rows not matching the filters.
For example: IF(CARDINALITY(array) > 0, array[1]))'

Test plan 
Added unit tests.
AggregateIfBenchmark result:

```
    Before:
    sum_if_without_rewrite :: 6514.329 cpu ms ::    0B peak memory :: in  1.5M,      0B,    230K/s,      0B/s :: out     1,      9B,       0/s,      1B/s
    sum_if_with_rewrite :: 5695.211 cpu ms ::    0B peak memory :: in  729K,      0B,    128K/s,      0B/s :: out     1,      9B,       0/s,      1B/s
    sum_filter :: 5771.367 cpu ms ::    0B peak memory :: in  729K,      0B,    126K/s,      0B/s :: out     1,      9B,       0/s,      1B/s

    After:
    sum_if_without_rewrite :: 6298.922 cpu ms :: 0B peak memory :: in 1.5M, 0B, 238K/s, 0B/s :: out 1, 9B, 0/s, 1B/s
    sum_if_with_rewrite :: 5820.618 cpu ms :: 0B peak memory :: in 729K, 0B, 125K/s, 0B/s :: out 1, 9B, 0/s, 1B/s
    sum_filter :: 5537.673 cpu ms :: 0B peak memory :: in 729K, 0B, 132K/s, 0B/s :: out 1, 9B, 0/s, 1B/s
```

```
== RELEASE NOTES ==

General Changes
* Disabled the Aggregation IF to FILTER rewrite rule by default. Updated the rule to stop unwrapping the IF expression instead Aggregation.
```