@yuanzhanhku 2021-08-30
> Can we add one test for each agg just to be sure? COUNT/ARBITRARY/SUM/MIN/MAX/MAX_BY/MIN_BY etc? In particular MAX_BY and MIN_BY will be good. I don't think they will match your pattern but good to test. This is for correctness test.

Done.